requirements.txt
requests
ping3
python-vlc
yt_dlp
textual